+++ CHANGELOG.md
@@ -24,7 +24,7 @@ All Notable changes to `cron` will be documented in this file
 
 ### Removed
 
-- None
+- **[BC Break]** `Scheduler::maxIterationCount` is no longer needed with the new implementation.
 
 ## [0.3.0] - 2021-12-31